+++ lib/value_cubit.dart
@@ -1,5 +1,5 @@
 library value_cubit;
 
 export 'src/cubit.dart';
-export 'src/helpers.dart';
+export 'src/extensions.dart';
 export 'src/states.dart';